+++ app/(user)/user/metaReDirect/page.tsx
@@ -1,7 +1,0 @@
-"use client";
-
-import PageComponent from "../../../../legacy-pages/user/metaReDirect/index";
-
-export default function Page() {
-  return <PageComponent />;
-}